Refactor ContainerTest to use VerifierImageFixture for Docker image management
